--- a/catanatron_core/catanatron/models/actions.py
+++ b/catanatron_core/catanatron/models/actions.py
@@ -52,13 +52,24 @@ def generate_playable_actions(state) -> List[Action]:
         return robber_possibilities(state, color)
     elif action_prompt == ActionPrompt.PLAY_TURN:
         if state.is_road_building:
-            actions = road_building_possibilities(state, color, False)
-        elif not player_has_rolled(state, color):
-            actions = [Action(color, ActionType.ROLL, None)]
-            if player_can_play_dev(state, color, "KNIGHT"):
-                actions.append(Action(color, ActionType.PLAY_KNIGHT_CARD, None))
+            return road_building_possibilities(state, color, False)
+        actions = []
+        # Allow playing dev cards before and after rolling
+        if player_can_play_dev(state, color, "YEAR_OF_PLENTY"):
+            actions.extend(year_of_plenty_possibilities(color, state.resource_freqdeck))
+        if player_can_play_dev(state, color, "MONOPOLY"):
+            actions.extend(monopoly_possibilities(color))
+        if player_can_play_dev(state, color, "KNIGHT"):
+            actions.append(Action(color, ActionType.PLAY_KNIGHT_CARD, None))
+        if (
+            player_can_play_dev(state, color, "ROAD_BUILDING")
+            and len(road_building_possibilities(state, color, False)) > 0
+        ):
+            actions.append(Action(color, ActionType.PLAY_ROAD_BUILDING, None))
+        if not player_has_rolled(state, color):
+            actions.append(Action(color, ActionType.ROLL, None))
         else:
-            actions = [Action(color, ActionType.END_TURN, None)]
+            actions.append(Action(color, ActionType.END_TURN, None))
             actions.extend(road_building_possibilities(state, color))
             actions.extend(settlement_possibilities(state, color))
             actions.extend(city_possibilities(state, color))
@@ -69,21 +80,6 @@ def generate_playable_actions(state) -> List[Action]:
             )
             if can_buy_dev_card:
                 actions.append(Action(color, ActionType.BUY_DEVELOPMENT_CARD, None))
-
-            # Play Dev Cards
-            if player_can_play_dev(state, color, "YEAR_OF_PLENTY"):
-                actions.extend(
-                    year_of_plenty_possibilities(color, state.resource_freqdeck)
-                )
-            if player_can_play_dev(state, color, "MONOPOLY"):
-                actions.extend(monopoly_possibilities(color))
-            if player_can_play_dev(state, color, "KNIGHT"):
-                actions.append(Action(color, ActionType.PLAY_KNIGHT_CARD, None))
-            if (
-                player_can_play_dev(state, color, "ROAD_BUILDING")
-                and len(road_building_possibilities(state, color, False)) > 0
-            ):
-                actions.append(Action(color, ActionType.PLAY_ROAD_BUILDING, None))
 
             # Trade
             actions.extend(maritime_trade_possibilities(state, color))