--- a/catanatron_core/catanatron/state.py
+++ b/catanatron_core/catanatron/state.py
@@ -76,6 +76,10 @@ PLAYER_INITIAL_STATE = {
     # de-normalized features (for performance since we think they are good features)
     "ACTUAL_VICTORY_POINTS": 0,
     "LONGEST_ROAD_LENGTH": 0,
+    "KNIGHT_OWNED_AT_START": False,
+    "MONOPOLY_OWNED_AT_START": False,
+    "YEAR_OF_PLENTY_OWNED_AT_START": False,
+    "ROAD_BUILDING_OWNED_AT_START": False,
 }
 for resource in RESOURCES:
     PLAYER_INITIAL_STATE[f"{resource}_IN_HAND"] = 0