@zarns 2024-11-12
Info on performance implications:
- catanatron-play -n 1000
  -  before: 43.4 sec
  - after: 44.1 sec
- catanatron-play --players=AB,R,W --num=100
  - before: 309.36 sec
  - after: 295.7 sec